feat(table): right-click acts on the selection
Right-clicking a row ignored the selection entirely: twenty rows
highlighted, menu acting on one. The bulk bar was the only multi-select
affordance, and the ambiguity was worst exactly where it matters — a
Delete pick that looks like it covers the selection and doesn't.

A right-click inside a multi-row selection now opens a selection menu:
header names the count, entries are the bulk verbs. A right-click outside
one drops the selection first, so the menu's scope always matches what is
highlighted. A single-row selection stays a row menu and keeps its
highlight — selection and menu already name the same row.

The verbs come from the same builders the BulkBar uses. That is the point:
a second set of "bulk" semantics that drifts from the first is how an
operator ends up deleting rows they thought they were only suspending. The
builders move out of App.tsx into components/bulkActions.ts with a
`bulkActionsFor` dispatcher so both surfaces resolve them the same way,
and `selectSelectionDegraded` moves into the store rather than being
recomputed per surface.

Single-row entries stay visible but disabled rather than vanishing — a
menu that silently loses "View details" reads as a bug, one that says
"one row only" explains itself.

The scope decision is `menuScopeFor`, pure and tested, rather than a
condition buried in a JSX handler.
